Gray scale function and color class.
Color class is nice logic, but for this implementation actually
slows us down since it doesn't utilize numpy vectorizations.

In the future, use this for complicated functions, but then replace
with numpy when possible. Next commit will be a more efficient
graying function.

Note for future: may want to condense starter code for processing image
into a function call.